requirements.txt
streamlit==1.36.0
streamlit-option-menu==0.3.13
pillow==10.4.0
joblib==1.4.2
numpy==2.0.1
pandas==2.2.2
scikit-learn==1.5.1
xgboost==2.1.0